adress requested changes from review
